refactor: Move src/Query.ts to src/Query/Query.ts
