+++ src/QueryRenderer.ts
@@ -10,7 +10,7 @@ import type { EventRef, MarkdownPostProcessorContext } from 'obsidian';
 import type { IQuery } from './IQuery';
 import { State } from './Cache';
 import { replaceTaskWithTasks } from './File';
-import { Query } from './Query';
+import { Query } from './Query/Query';
 import type { GroupHeading } from './Query/GroupHeading';
 import { TaskModal } from './TaskModal';
 import type { Events } from './Events';
